Fixes from initial validation with hardware. First example working

+++ examples/qwiic_vl53l5cx_ex6_settings.py
@@ -57,7 +57,7 @@ def runExample():
 
 	myVL53L5CX.set_resolution(8*8) # enable all 64 pads
 	image_resolution = myVL53L5CX.get_resolution()  # Query sensor for current resolution - either 4x4 or 8x8
-	image_width = sqrt(image_resolution) # Calculate printing width
+	image_width = int(sqrt(image_resolution)) # Calculate printing width
 	
 	# Set the ranging mode
 	if myVL53L5CX.set_ranging_mode(myVL53L5CX.kRangingModeAutonomous) != myVL53L5CX.kStatusOK:
