+++ examples/qwiic_vl53l5cx_ex8_set_address.py
@@ -68,7 +68,7 @@ def runExample():
 	myVL53L5CX.set_resolution(8*8) # enable all 64 pads
 	image_resolution = myVL53L5CX.get_resolution()  # Query sensor for current resolution - either 4x4 or 8x8
 
-	image_width = sqrt(image_resolution) # Calculate printing width
+	image_width = int(sqrt(image_resolution)) # Calculate printing width
 	myVL53L5CX.start_ranging()
 
 	while True:
